requirements.txt
PyYAML==3.12
argparse==1.2.1
fitbit==0.2.3
numpy==1.11.2
oauthlib==2.0.0
pandas==0.19.1
python-dateutil==2.4.2
python-gflags==3.1.0
pytz==2016.7
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
wsgiref==0.1.2
